refactor trainer rebase
