+++ paddleformers/transformers/context_parallel_utils.py
@@ -26,6 +26,7 @@
 
 
 import paddle
+from paddle.distributed.auto_parallel.ring_attention import shard_seq_load_balance
 from paddle.distributed.fleet import fleet
 
 
@@ -59,6 +60,25 @@ def split_inputs_sequence_dim_load_balance(inputs, rank=None, degree=None):
         res = []
         for tensor in inputs:
             res.append(do_split_sequence_dim_load_balance(tensor, rank, degree))
+    else:
+        raise ValueError(f"the inputs should be a list or a dict, but is type: {type(inputs)}")
+    return res
+
+
+def auto_split_sequence_dim_load_balance(inputs):
+    """
+    for auto_parallel mode
+    """
+    if isinstance(inputs, paddle.Tensor):
+        return shard_seq_load_balance(inputs, 1)
+    elif isinstance(inputs, dict):
+        res = {}
+        for k, tensor in inputs.items():
+            res[k] = shard_seq_load_balance(tensor, 1)
+    elif isinstance(inputs, list):
+        res = []
+        for tensor in inputs:
+            res.append(shard_seq_load_balance(tensor, 1))
     else:
         raise ValueError(f"the inputs should be a list or a dict, but is type: {type(inputs)}")
     return res
